+++ CHANGES.rst
@@ -16,6 +16,7 @@ develop
 - ui: add foundation for page transitions from codrops
 - hiveeyes: receive messages via MQTT and store data points into InfluxDB
 - app: use configuration file instead of hardcoded configuration values
+- storage: add support for InfluxDB 0.9
 
 
 2015-05-21 0.2.2